+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/RightPanel.tsx
@@ -14,7 +14,7 @@ import { useTreeRef } from "./treeview/provider/TreeRefProvider";
 import { TreeView } from "./treeview/TreeView";
 import { useRehydrate } from "@lib/store/useTemplateStore";
 
-import { NextActions } from "@formBuilder/[id]/edit/components/logic/NextActions";
+import { SelectNextAction } from "./logic/SelectNextAction";
 import { useGroupStore } from "./treeview/store/useGroupStore";
 
 const TabButton = ({
@@ -188,7 +188,7 @@ export const RightPanel = ({ id }: { id: string }) => {
                       </Tab.Panel>
                       <Tab.Panel>
                         <div className="m-0 mt-1 w-full border-t-1 border-slate-200 p-10">
-                          {item && <NextActions item={item} />}
+                          {item && <SelectNextAction item={item} />}
                         </div>
                       </Tab.Panel>
                     </Tab.Panels>
